Introduce config flag for N26 to only import processed transactions
Because the values that we get from the (not official) N26 API change from time to time, people often end up with duplicated transactions. There as already been effort in avoiding that (#16) but it is still happening.

This PR will introduce a new config flag for the [N26 Dumper](https://github.com/schurig/ynab-bank-importer/wiki/n26) called `skip_pending_transactions`. This setting will be set to `false` by default. When set to true, all transactions that are still pending _(= not processed by N26)_ won't be imported by the script until they are processed. This will prevent duplicate entries but comes with the tradeoff that it might take a few days for transactions to show up in your YNAB.